+++ lib/src/common/model/partial_payment/order_cancel_result.dart
@@ -1,0 +1,9 @@
+class OrderCancelResult {
+  final Map<String, dynamic> orderCancelResponseBody;
+  Map<String, dynamic>? updatedPaymentMethodsResponseBody;
+
+  OrderCancelResult({
+    required this.orderCancelResponseBody,
+    this.updatedPaymentMethodsResponseBody,
+  });
+}